==> analysis/quantifier-comparisons/08-genes-detected-comparison.Rmd <==
---
title: "Analysis of Genes detected between Tools"
author: "Ally Hawkins for CCDL"
output: 
  html_notebook:
    toc: true
    toc_float: true
---

This is a quick comparison of the genes that are detected when using cellranger or when using alevin-fry. 
Here, we have alevin-fry that has been run in two different resolution modes, `cr-like` or `cr-like-em`. 

## Setup

```{r}
set.seed(1234)
library(magrittr)
library(ggplot2)
library(ggupset)
library(ggrepel)
library(SingleCellExperiment)
library(clusterProfiler)
library(org.Hs.eg.db)
```

```{r}
# path to results files with sces and qc dataframes from running benchrmarking_generate_qc_df.R
base_dir <- here::here()
file_dir <- file.path(base_dir, "data", "cr-like-test", "results")

# qc files 
rowdata_df_file <- file.path(file_dir, "rowdata_qc.tsv")
coldata_df_file <- file.path(file_dir, "coldata_qc.tsv")
quant_info_file <- file.path(file_dir, "quant_info.tsv")

# sce objects
cr_like_file <- file.path(file_dir, "splici_salign_cr-like_sces.rds")
cr_like_em_file <- file.path(file_dir, "splici_salign_cr-like-em_sces.rds")
cellranger_file <- file.path(file_dir, "cellranger_sces.rds")
```

```{r}
# read in rowdata_df containing information about each gene
rowdata_df <- readr::read_tsv(rowdata_df_file)
coldata_df <- readr::read_tsv(coldata_df_file)
quant_info <- readr::read_tsv(quant_info_file)
```

```{r}
# read in sces
cr_like_sce <- readr::read_rds(cr_like_file)
cr_like_em_sce <- readr::read_rds(cr_like_em_file)
cellranger_sce <- readr::read_rds(cellranger_file)
```


```{r}
# join coldata with quant_info 
alevin_fry_tools <- c("splici_salign_cr-like", "splici_salign_cr-like-em")
# merge coldata df with quant_info
coldata_info_df <- coldata_df %>%
  # rename tool to be either cellranger or alevin-fry
  dplyr::mutate(tool = ifelse(tool %in% alevin_fry_tools, "alevin-fry", "cellranger")) %>%
  dplyr::left_join(quant_info,
                   by = c("tool" = "tool", 
                          "quant_id" = "quant_dir")) %>%
  # rename the not_alevin to cellranger for plotting purposes
  dplyr::mutate(alevin_resolution = ifelse(alevin_resolution == "not_alevin", "cellranger", alevin_resolution))

# modify quant info to have alevin resolution as cellranger rather than not_alevin
quant_info <- quant_info %>%
  dplyr::mutate(alevin_resolution = ifelse(alevin_resolution == "not_alevin", "cellranger", alevin_resolution))
```


```{r}
# filter for common cells across all samples 
cell_counts <- coldata_info_df %>%  
  dplyr::count(cell_id, sample)

common_cells <- cell_counts %>%
  dplyr::filter(n == 3) %>%
  dplyr::pull(cell_id)
```

```{r}
# make a function to filter sces by subset of cells and re-calculate feature stats
filter_sce <- function(sce, cells){
  cells_to_keep <- colnames(sce) %in% cells
  SummarizedExperiment::rowData(sce) <- NULL
  sce[, cells_to_keep] %>%
    scater::addPerFeatureQC()
}
```


```{r}
# filter all sces together 
all_sces <- list(cr_like_sce,
                 cr_like_em_sce,
                 cellranger_sce)
names(all_sces) <- c('cr-like', 'cr-like-em', 'cellranger')

sces_filtered <- all_sces %>%
  purrr::map(
    ~ purrr::map(.x, filter_sce, cells = common_cells)
  )

# grab rowdata from filtered sces 
rowdata_filtered <- sces_filtered %>% 
  purrr::map_df(
    ~ purrr::map_df(.x, scpcaTools::rowdata_to_df, .id = "quant_dir"),
    .id = "alevin_resolution"
  ) %>%
  dplyr::left_join(quant_info)
```

## Differences in Gene detection between Alevin-fry and Cellranger

### Overlap of Genes 

As another comparison, let's look at the actual genes that are identified in each of the tools. 
Let's see how many are lost when we use alevin-fry over cellranger and what types of genes they are. 
Here, we will consider all genes that have the capability to be detected in alevin-fry and cellranger, regardless of if they are expressed or not. 
We want to restrict the analysis to only genes detected in both tools to avoid any differences based on ensembl versions as we used ensembl v100 with cellranger and v103 with alevin-fry.

```{r}
# get genes common in all tools
common_genes <- rowdata_filtered %>%
  dplyr::select(gene_id, tool) %>%
  dplyr::distinct() %>%
  dplyr::group_by(gene_id) %>%
  dplyr::tally() %>%
  # filter for genes found in both cellranger and alevin-fry
  dplyr::filter(n == 2) %>%
  dplyr::pull(gene_id)

# filter rowdata_df to only include genes found in all tools 
rowdata_filtered <- rowdata_filtered %>% 
  dplyr::filter(gene_id %in% common_genes)
  
```


```{r}
# pull out genes that are detected and expressed in each of the tools
cellranger_genes <- rowdata_filtered %>%
  dplyr::filter(tool == "cellranger" & mean > 0 & detected > 0) %>%
  dplyr::pull(gene_id) %>%
  unique()

cr_like_genes <- rowdata_filtered %>%
  dplyr::filter(alevin_resolution == "cr-like" & mean > 0 & detected > 0) %>%
  dplyr::pull(gene_id) %>%
  unique()

cr_em_genes <- rowdata_filtered %>%
  dplyr::filter(alevin_resolution == "cr-like-em" & mean > 0 & detected > 0) %>%
  dplyr::pull(gene_id) %>%
  unique()
```


```{r}
# get intersection gene lists 
# look at intersection between cellranger and alevin-fry cr-like
cellranger_cr_intersect <- intersect(cellranger_genes, cr_like_genes)

# look at intersection between cellranger and alevin-fry cr-like-em
cellranger_cr_em_intersect <- intersect(cellranger_genes, cr_em_genes)

# look at intersection between alevin-fry cr-like and alevin-fry cr-like-em
cr_cr_em_intersect <- intersect(cr_like_genes, cr_em_genes)
```

```{r}
# filter rowdata to only include genes with mean > 0 and detected > 0 for rest of comparisons
rowdata_filtered <- rowdata_filtered %>%
  dplyr::filter(mean > 0 & detected > 0)
```


```{r}
# get unique gene lists by getting genes that overlap with all other tools first and then the set difference with the list of genes for that tool
all_cellranger_overlap <- rowdata_filtered %>%
  dplyr::filter(gene_id %in% cellranger_genes & tool =="alevin-fry") %>%
  dplyr::pull(gene_id) %>%
  unique()
cellranger_unique <- setdiff(cellranger_genes, all_cellranger_overlap)

cr_overlap <- rowdata_filtered %>%
  dplyr::filter(gene_id %in% cr_like_genes & alevin_resolution %in% c("cr-like-em", "not_alevin")) %>%
  dplyr::pull(gene_id) %>%
  unique()
cr_unique <- setdiff(cr_like_genes, cr_overlap)

cr_em_overlap <- rowdata_filtered %>%
  dplyr::filter(gene_id %in% cr_em_genes & alevin_resolution %in% c("cr-like", "not_alevin")) %>%
  dplyr::pull(gene_id) %>%
  unique()
cr_em_unique <- setdiff(cr_em_genes, cr_em_overlap)
```


```{r}
# visualize overlap between gene sets based on tool

gene_detect_df <- rowdata_filtered %>%
  group_by(gene_id) %>%
  dplyr::summarise(tools_detected = list(unique(alevin_resolution)))

ggplot(gene_detect_df, aes(x = tools_detected)) +
  geom_bar() +
  scale_x_upset(n_intersections = 6)
```

When looking at shared cells only and limiting to only genes that are found in both indices, we see that the majority of genes are shared across all tools. 
Alevin-fry `cr-like-em` contains the most unique genes with cellranger the second most unique genes although the numbers here are quite small at 318 and 117 genes respectively. 

Perhaps the genes that are different are lowly expressed genes and it wouldn't matter anyway. 

If we now look at the gene expression for those genes that are unique to each tool, how do they compare to the gene expression for those genes that are shared?


```{r}
# calculate mean expression of log transformed counts matrix 
log_row_means <- function(sce){
   rowMeans(log1p(counts(sce))) %>%
    as.data.frame() %>%
    tibble::rownames_to_column("gene_id") %>%
    dplyr::rename("mean_log_expression" = ".")
}

# get list of row mean data frames 
mean_exp_list <- sces_filtered %>%
  purrr::map(
    ~ purrr::map(.x, log_row_means)
  )

# join together into larger data frame with quant_dir and alevin_resolution as unique ID's
mean_exp_df <- mean_exp_list %>%
  purrr::map_df(
    ~ purrr::map_df(.x, rbind, .id = "quant_dir"), 
    .id = "alevin_resolution"
  ) %>%
  dplyr::left_join(quant_info)

# join back with quant_info to get sample information 
```


Now, we want to classify each gene as being shared by all of the tools or unique. 
We will label genes found in all tools as shared, genes found in 2 of 3 tools as shared and then genes found in only that one tool as unique. 
```{r}
# add a column to classify as shared or unique to that tool
mean_exp_df <- mean_exp_df %>%
  # filter to only include genes that are common and have mean > 0 and detected > 0 
  dplyr::filter(gene_id %in% rowdata_filtered$gene_id) %>%
  dplyr::mutate(gene_status = dplyr::case_when(gene_id %in% intersect(cellranger_cr_intersect, cellranger_cr_em_intersect) ~ "common",
  gene_id %in% c(cellranger_cr_intersect, cellranger_cr_em_intersect, cr_cr_em_intersect) ~ "shared", 
  TRUE ~ "unique")) 
```


```{r fig.height = 5, fig.width=10}
# make a plot to look at the mean gene expression of shared vs. unique genes for each tool
ggplot(mean_exp_df, aes(x = alevin_resolution, y =mean_log_expression , fill = gene_status)) + 
  geom_boxplot() + 
  facet_wrap(~sample, nrow = 2) + 
  coord_cartesian(ylim = c(0,1))
```

As expected, across the board, it looks like the unique genes are those with lower mean gene expression than those genes that are shared across all tools. 
The genes that are lost are probably not contributing very much to downstream results. 

Let's look at the few unique genes in alevin-fry that are higher expressed. 
There don't appear to be any in cellranger. 
 
```{r}
cellranger_unique_high <- mean_exp_df %>%
  dplyr::filter(alevin_resolution == "cellranger" & 
                  gene_status == "unique" & 
                  mean_log_expression > 0.1) %>%
  dplyr::pull(gene_id)

rowdata_filtered %>%
  dplyr::filter(gene_id %in% cellranger_unique_high)
```
 
```{r}
cr_like_em_unique_high <- mean_exp_df %>%
  dplyr::filter(alevin_resolution == "cr-like-em" & 
                  gene_status == "unique" & 
                  mean_log_expression > 0.10) %>%
  dplyr::pull(gene_id)

rowdata_filtered %>%
  dplyr::filter(gene_id %in% cr_like_em_unique_high) %>%
  dplyr::arrange(desc(mean))
```

These genes that are found in cr-like-em include the following: 

- ENSG00000214736 - TOMM6
- ENSG00000285839 - KTI12
- ENSG00000248167 - TRIM39-RPP21
- ENSG00000285508 - ALO34330.1

### Correlation of gene expression between Cellranger and Alevin-fry

We can also look at the correlation of gene expression across these tools and use a linear regression model to see how similar the mean gene expression determined by cellranger is to each of the alevin-fry methods. 
Here, I am using a linear regression model for each sample and grabbing the residuals for each gene. 
The residuals will help show us if alevin-fry is over or underestimating gene counts in comparison to cellranger.

```{r}
# create separate columns for mean log expression for each tool
mean_exp_df_spread <- mean_exp_df %>%
  tidyr::pivot_wider(id_cols = c(gene_id, sample),
                     names_from = alevin_resolution,
                     values_from = mean_log_expression)

# use a linear regression model to compare gene expression across tools 
# grab the residuals for each gene and add to the original dataframe
residuals_df <- mean_exp_df_spread %>%
  dplyr::group_by(sample) %>%
  dplyr::mutate(
    cellranger_cr_like = lm(`cr-like` ~ cellranger )$residuals,
    cellranger_cr_like_em = lm(`cr-like-em` ~ cellranger)$residuals,
    cr_cr_like_em = lm(`cr-like-em` ~ `cr-like`)$residuals
  ) %>%
  # condense all residuals for each tool comparison into one column for easy plotting
  tidyr::pivot_longer(cols = c("cellranger_cr_like", "cellranger_cr_like_em", "cr_cr_like_em"),
                      names_to = "tool_comparisons", 
                      values_to = "residuals") %>%
 dplyr::rename("cr_mean_log_expression" = "cellranger",
               "crlike_mean_log_expression" = "cr-like",
               "crlikeEM_mean_log_expression" = "cr-like-em")
```

```{r fig.height=10, fig.width=10}
cellranger_comparisons <- residuals_df %>%
  dplyr::filter(tool_comparisons %in% c("cellranger_cr_like", "cellranger_cr_like_em"))

ggplot(cellranger_comparisons, aes(x = cr_mean_log_expression, y = residuals)) + 
  geom_point(size = 0.5) + 
  facet_grid(sample ~ tool_comparisons)+ 
  # label genes that have abs(residual) > 1 to see which genes are consistently labeled across samples
  geom_text_repel(data = cellranger_comparisons %>%
              dplyr::filter(abs(residuals) > 1), 
            mapping = aes(x = cr_mean_log_expression, y = residuals, label = gene_id),
            size = 3, max.overlaps = 20)
```

Interestingly it looks like the single-nuclei samples (220 and 221) have fewer differences between the two tools and show a lot more consistency between the resolution modes. 
It also looks like regardless of alevin-fry resolution strategy, the same genes will be under or over represented. 
We also see an increase in positive residuals in the `cr-like-em` to `cellranger` comparison, suggesting that the model is underestimating the observed values seen in `cr-like-em` and the observed expression for these genes is higher than anticipated. 
On the contrary, `cr-like` has a few outliers with negative residuals meaning it is undercounting gene expression in compared to cellranger. 
And finally, the residuals appear to move closer to 0 as gene expression increases.  

```{r}
# print out a table with all the gene names shown above 
gene_symbols_df <- rowdata_df %>%
  dplyr::select(gene_id, Symbol)

cellranger_comparisons %>%
  dplyr::left_join(gene_symbols_df) %>%
  dplyr::distinct() %>%
  tidyr::drop_na() %>%
  dplyr::filter(abs(residuals) > 1) %>%
  dplyr::arrange(desc(residuals)) %>%
  dplyr::select(sample, Symbol, gene_id, residuals, tool_comparisons, cr_mean_log_expression, crlike_mean_log_expression, crlikeEM_mean_log_expression)
```

If you scroll through the list of genes that have abs(residuals) > 1 there are a lot of ribosomal genes that tend to be under counted by alevin-fry. 
The genes that are over counted by alevin-fry are a mix of coding genes and pseudogenes. 
We also see that `cr-like-em` tends to have positive residuals and over count and `cr-like` has more negative residuals and undercount.

Let's look at the comparison of `cr-like` to `cr-like-em`. 

```{r fig.height=10, fig.width=5}
crlike_comparisons <- residuals_df %>%
  dplyr::filter(tool_comparisons == "cr_cr_like_em")

ggplot(crlike_comparisons, aes(x = crlike_mean_log_expression, y = residuals)) + 
  geom_point(size = 0.5) + 
  facet_grid(sample ~ tool_comparisons)+ 
  # label genes that have abs(residual) > 1 to see which genes are consistently labeled across samples
  geom_text_repel(data = crlike_comparisons %>%
              dplyr::filter(abs(residuals) > 1), 
            mapping = aes(x = crlike_mean_log_expression, y = residuals, label = gene_id),
            size = 3, max.overlaps = 20)
```


```{r}
# print out a table with all the gene names shown above 
crlike_comparisons %>%
  dplyr::left_join(gene_symbols_df) %>%
  dplyr::distinct() %>%
  tidyr::drop_na() %>%
  dplyr::filter(abs(residuals) > 1) %>%
  dplyr::arrange(desc(residuals)) %>%
  dplyr::select(sample, Symbol, gene_id, residuals, tool_comparisons, cr_mean_log_expression, crlike_mean_log_expression, crlikeEM_mean_log_expression)
```
Here you can see that most of the genes are over counted rather than undercounted by `cr-like-em` in comparison to `cr-like` which makes sense given `cr-like-em` resolves multi mapped genes. 
Again, we see mostly ribosomal genes are affected. 

### Gene Ontology

Let's make sure that we aren't losing any genes that we might care about and would be important pathways in pediatric oncology. 

First we need to identify the genes that are unique to each of the tools and would be lost if we weren't using that tool. 

```{r}
# join back with initial rowdata_df to get symbol to use for gene set enrichment analysis 
cellranger_rowdata <- rowdata_df %>%
  dplyr::filter(tool == "cellranger") %>%
  dplyr::select(gene_id, Symbol)

cellranger_lost <- rowdata_filtered %>%
  dplyr::filter(gene_id %in% cellranger_unique) %>%
  dplyr::left_join(cellranger_rowdata) %>%
  dplyr::arrange(desc(mean)) %>%
  dplyr::select(gene_id, mean, Symbol, detected, sample) %>%
  unique()

cr_em_rowdata <- rowdata_df %>%
  dplyr::filter(tool == "splici_salign_cr-like-em") %>%
  dplyr::select(gene_id, Symbol)

cr_em_lost <- rowdata_filtered %>%
  dplyr::filter(!(gene_id %in% cellranger_cr_em_intersect)) %>%
  dplyr::left_join(cr_em_rowdata) %>%
  dplyr::arrange(desc(mean)) %>%
  dplyr::select(tool, gene_id, Symbol, mean, detected, sample) %>%
  unique()

```

Since we are specifically interested in looking at what we would lose if we were to use alevin-fry over cellranger, let's focus on the genes lost in cellranger. 

```{r}
cellranger_lost
summary(cellranger_lost$mean)
summary(cellranger_lost$detected)
```


Notice that because we are looking at 4 different samples that have been run with cellranger, each gene is represented 4 times. 
It looks like for the most part they are lowly expressed other than the one gene, PRRC2B, but is found in the majority of cells.

These genes probably do not have an impact on downstream results if we are missing them, but let's be sure that we aren't missing out on any major pathways we would care about. 

Here, I am doing over representation analysis with the lost genes from not using cellranger as the target gene list and the background gene list being all genes identified in our tools. 

```{r}
cellranger_lost_gene_list <- cellranger_lost$Symbol %>%
  unique()
background_genes <- rowdata_df$Symbol %>%
  unique()
```

```{r}
# look to see what genes are found in cellranger and not found in alevin-fry
cellranger_lost_gene_list
```

```{r}
go_ora_results <- enrichGO(gene = cellranger_lost_gene_list,
                           universe = background_genes,
                           keyType = "SYMBOL",
                           OrgDb = org.Hs.eg.db,
                           ont = "BP",
                           pAdjustMethod = "BH",
                           pvalueCutoff = 0.00001)
```


```{r}
go_results <- go_ora_results@result %>%
  as.data.frame() %>%
  dplyr::filter(p.adjust < 0.00001)
go_results
```

```{r fig.width = 5}
enrichplot::dotplot(go_ora_results)
```


It looks like we are getting enrichment all of the same similar pathways involved in sensory perception and chemical stimulus. 

What about the type of genes that are unique to alevin-fry `cr-like-em`. 

```{r}
cr_em_lost_genes <- cr_em_lost$gene_id %>%
  unique()
background_genes <- rowdata_df$gene_id %>%
  unique()
```

```{r}
go_ora_results <- enrichGO(gene = cr_em_lost_genes,
                           universe = background_genes,
                           keyType = "ENSEMBL",
                           OrgDb = org.Hs.eg.db,
                           ont = "BP",
                           pAdjustMethod = "BH",
                           pvalueCutoff = 0.00001)
```


```{r}
go_results <- go_ora_results@result %>%
  as.data.frame() %>%
  dplyr::filter(p.adjust < 0.00001)
go_results
```


```{r fig.width=5}
enrichplot::dotplot(go_ora_results)
```

Now if we look at the types of genes that are identified uniquely in Alevin-fry `cr-like-em` as opposed to Cellranger or Alevin-fry `cr-like`, we see similar sensory and chemical stimulus pathways as well as some snRNA processing. 

```{r}
sessioninfo::session_info()
```
